Move clean command to common build_ext module.
