+++ build_ext/utils.py
@@ -16,7 +16,27 @@ from __future__ import print_function, division, absolute_import
 import fnmatch
 import os
 
+from glob import glob
+
 from distutils import cmd
+from distutils.command.clean import clean as _clean
+from distutils.dir_util import remove_tree
+
+
+class clean(_clean):
+    def initialize_options(self):
+        self.egg_base = None
+        _clean.initialize_options(self)
+
+    def finalize_options(self):
+        self.set_undefined_options('egg_info', ('egg_base', 'egg_base'))
+        _clean.finalize_options(self)
+
+    def run(self):
+        if self.all:
+            for f in glob(os.path.join(self.egg_base, '*.egg-info')):
+                remove_tree(f, dry_run=self.dry_run)
+        _clean.run(self)
 
 
 def memoize(f):
